+++ src/client/guildMember.ts
@@ -1,0 +1,10 @@
+import { GuildMember } from "discord.js";
+
+declare module "discord.js" {
+  interface GuildMember {
+    createdTimestamp: number;
+    reason: string;
+  }
+}
+
+export { GuildMember };