+++ src/client/index.ts
@@ -1,5 +1,6 @@
 export { ExtendedClient as Client } from "./client";
 export { ExtendedEmbed as MessageEmbed } from "./embed";
+export * from "./guildMember";
 export * from "./message";
 export * from "./automation";
 export * from "./interaction";