+++ src/events/guildMemberRemove.ts
@@ -1,10 +1,15 @@
-import { Client } from "@client";
-import { GuildMember } from "discord.js";
+import { Client, GuildMember } from "@client";
 import { Event } from "@interfaces";
 
 export const event: Event = {
 	name: "guildMemberRemove",
 	run: async (client: Client, member: GuildMember) => {
+
+		//! do not remove, 'force' member to be casted (break if removed)
+		let _ = (member as GuildMember) instanceof GuildMember;
+
+		member.createdTimestamp = new Date().getTime();
+		member.reason = "removed";
 		client.storeAction("guildMemberUpdate", member);
 
 		const updateChannel: string = client.config.discords.filter((s) => s.id === member.guild.id).pop()