+++ src/events/messageCreate.ts
@@ -7,11 +7,14 @@ import { getSubmissionsChannels } from "@helpers/channels";
 export const event: Event = {
 	name: "messageCreate",
 	run: async (client: Client, message: Message) => {
-		if (message.author.bot) return;
-		client.storeAction("message", message);
-		
 		//! do not remove, 'force' message to be casted (break if removed)
 		let _ = (message as Message) instanceof Message;
+		
+		let m = Object.assign({}, message); // loose reference to message: create unique instance of the message for the logger (ask @Juknum)
+		m.isDeleted = false;
+		client.storeAction("message", m);
+		
+		if (message.author.bot) return;
 
 		// old commands handler (remove for release)
 		if (message.content.startsWith(client.tokens.prefix)) {